Re-Write of RK4
Seperated out the step into a function which uses a for loop to calculate the intermediate values.

This makes the program smaller and a little easier to read (I think). Also one can call the function and just step the integrator forward without capturing the results at all (if they want).